+++ plugins/contenttraker/src/main.ts
@@ -1,0 +1,3 @@
+import { connectContentTrakerMcpServer } from "./server.js";
+
+await connectContentTrakerMcpServer();